chore: checker の型を types に移動
