+++ app/features/checker/check.ts
@@ -1,36 +1,9 @@
+import type {
+  StaticChecker,
+  DynamicChecker,
+  CheckResult,
+} from "~/types/checker";
 import { runCode } from "../code-runner/codeRunner";
-
-/**
- * 静的テストは、コードを受け取ってチェックする
- */
-interface StaticChecker {
-  description: string;
-  check: (code: string) => boolean;
-
-  // ユーザに対して表示されるメッセージ
-  message?: string;
-}
-
-/**
- * 動的テストは、出力を受け取ってチェックする
- */
-interface DynamicChecker {
-  description: string;
-  check: (out: string) => boolean;
-
-  // ユーザに対して表示されるメッセージ
-  message?: string;
-}
-
-type Checker = StaticChecker | DynamicChecker;
-
-/**
- * ユーザの入力したコードに対して行われるチェックの結果を表すオブジェクト。
- */
-interface CheckResult {
-  status: "success" | "failed";
-  failedChecker?: Checker;
-}
 
 /**
  * ユーザの入力に対し、テストを行う。
